requirements_rag.txt
# Core dependencies
python-dotenv>=1.0.0
supabase-py>=2.4.0
anthropic>=0.25.0

# RAG & Embeddings
sentence-transformers>=2.5.0
numpy>=1.24.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# PDF processing
PyMuPDF>=1.24.0

# Google Drive (optional)
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# Browser automation (optional, para coleta)
playwright>=1.40.0

# Utilities
beautifulsoup4>=4.12.0
requests>=2.31.0
